stream-app/requirements.txt
Flask==2.1.0
gunicorn==20.1.0
web3
google-cloud-pubsub
asyncio
fastavro
python-dotenv
google-cloud-dataflow-client
google-api-python-client
google-api-core